Add secure cipher suite list to extra args for kubernetes components
