+++ pkg/clusterapi/extraargs.go
@@ -2,8 +2,10 @@ package clusterapi
 
 import (
 	"fmt"
+	"strings"
 
 	"github.com/aws/eks-anywhere/pkg/api/v1alpha1"
+	"github.com/aws/eks-anywhere/pkg/crypto"
 	"github.com/aws/eks-anywhere/pkg/logger"
 	"github.com/aws/eks-anywhere/pkg/templater"
 )
@@ -45,6 +47,14 @@ func PodIAMAuthExtraArgs(podIAMConfig *v1alpha1.PodIAMConfig) ExtraArgs {
 	}
 	args := ExtraArgs{}
 	args.AddIfNotEmpty("service-account-issuer", podIAMConfig.ServiceAccountIssuer)
+	return args
+}
+
+// We don't need to add these once the Kubernetes components default to using the secure cipher suites
+func SecureTlsCipherSuitesExtraArgs() ExtraArgs {
+	args := ExtraArgs{}
+	cipherSuitesString := strings.Join(crypto.SecureCipherSuiteNames(), ",")
+	args.AddIfNotEmpty("tls-cipher-suites", cipherSuitesString)
 	return args
 }
