+++ pkg/providers/docker/config/template-cp.yaml
@@ -121,9 +121,15 @@ spec:
         extraArgs:
           enable-hostpath-provisioner: "true"
           profiling: "false"
+{{- if .controllermanagerExtraArgs }}
+{{ .controllermanagerExtraArgs.ToYaml | indent 10 }}
+{{- end }}
       scheduler:
         extraArgs:
           profiling: "false"
+{{- if .schedulerExtraArgs }}
+{{ .schedulerExtraArgs.ToYaml | indent 10 }}
+{{- end }}
     files:
     - content: |
 {{ .auditPolicy | indent 8 }}
@@ -172,6 +178,9 @@ spec:
         kubeletExtraArgs:
           cgroup-driver: cgroupfs
           eviction-hard: nodefs.available<0%,nodefs.inodesFree<0%,imagefs.available<0%
+{{- if .kubeletExtraArgs }}
+{{ .kubeletExtraArgs.ToYaml | indent 10 }}
+{{- end }}
 {{- if .controlPlaneTaints }}
         taints: {{ range .controlPlaneTaints}}
           - key: {{ .Key }}
@@ -190,6 +199,9 @@ spec:
         kubeletExtraArgs:
           cgroup-driver: cgroupfs
           eviction-hard: nodefs.available<0%,nodefs.inodesFree<0%,imagefs.available<0%
+{{- if .kubeletExtraArgs }}
+{{ .kubeletExtraArgs.ToYaml | indent 10 }}
+{{- end }}
 {{- if .controlPlaneTaints }}
         taints: {{ range .controlPlaneTaints}}
           - key: {{ .Key }}
